Simulator keyboard and localization settings, multiple test filters, streaming test events instead of writing them at test suite finish

+++ FBSimulatorControl/HID/FBSimulatorIndigoHID.h
@@ -1,8 +1,10 @@
 /**
- * Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
  *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
  */
 
 #import <CoreGraphics/CoreGraphics.h>
@@ -13,7 +15,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  An Enumeration for the Direction of the Event.
  */
-typedef NS_ENUM(int, FBSimulatorHIDDirection) {
+typedef NS_ENUM(NSUInteger, FBSimulatorHIDDirection) {
   FBSimulatorHIDDirectionDown = 1,
   FBSimulatorHIDDirectionUp = 2,
 };
@@ -21,7 +23,7 @@ typedef NS_ENUM(int, FBSimulatorHIDDirection) {
 /**
  An Enumeration Representing a button press.
  */
-typedef NS_ENUM(int, FBSimulatorHIDButton) {
+typedef NS_ENUM(NSUInteger, FBSimulatorHIDButton) {
   FBSimulatorHIDButtonApplePay = 1,
   FBSimulatorHIDButtonHomeButton = 2,
   FBSimulatorHIDButtonLock = 3,
